requirements.txt
# Chatnary Python Backend - Optimized Requirements

# FastAPI Core (stable versions)
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
python-multipart>=0.0.5

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Database
motor>=3.3.0
pymongo>=4.6.0

# AI & RAG Dependencies (locked for stability)
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0
openai>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
google-generativeai>=0.3.0

# Document Processing
pypdf>=3.17.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Utilities
aiofiles>=23.0.0
python-dateutil>=2.8.0
typing-extensions>=4.8.0

# HTTP Client for testing
httpx>=0.24.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0